Revert "Teste linux 2"

+++ lib/General.h
@@ -17,10 +17,12 @@
 #define QTD_ORDENACOES 4 //Expansivel para analisar mais de uma ordenacao
 
 
-unsigned long int TAM; //Pode ser tanto o padrao como de um arquivo
-unsigned long int MEIO; //Metade de TAM
+int TAM; //Pode ser tanto o padrao como de um arquivo
+int MEIO; //Metade de TAM
 
-unsigned long int *original; // Vetor default para caso base da analise
+int *original; // Vetor default para caso base da analise
+
+//#include "../src/sort.cpp"
 
 #endif
